Internalize Nimbus JwtDecoder Builder
Issue: gh-6010
